1.21 port.

+++ src/main/java/org/zeith/improvableskills/custom/skills/SkillEnchanter.java
@@ -4,7 +4,7 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.inventory.EnchantmentMenu;
 import net.minecraft.world.level.storage.loot.BuiltInLootTables;
 import net.minecraft.world.phys.AABB;
-import net.minecraftforge.event.enchanting.EnchantmentLevelSetEvent;
+import net.neoforged.neoforge.event.enchanting.EnchantmentLevelSetEvent;
 import org.zeith.improvableskills.api.registry.PlayerSkillBase;
 import org.zeith.improvableskills.data.PlayerDataManager;
 
@@ -32,7 +32,7 @@ public class SkillEnchanter
 				/* Check that the item is equal by memory reference. Allows to see who is actually calling the event. Little hack ;) */
 				if(e.getItem() == ench.enchantSlots.getItem(0))
 				{
-					int enchanter = PlayerDataManager.handleDataSafely(p, data -> data.isSkillActive(this) ? data.getSkillLevel(this) : 0, 0).intValue();
+					int enchanter = PlayerDataManager.handleDataSafely(p, data -> data.isSkillActive(this) ? (int) data.getSkillLevel(this) : 0, 0).intValue();
 					if(enchanter > 0 && e.getEnchantLevel() != 0)
 						e.setEnchantLevel(Math.max(1, e.getEnchantLevel() - enchanter / 4));
 					return;
